# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HH Education API
  version: 0.1.0
  description: API documentation for the current backend implementation.
servers:
  - url: http://localhost:4000
    description: Local development server
...
paths:
...
  /admin/assessments:
    get:
      tags:
        - Assessments
      summary: List admin assessments
      description: Admin and teacher list non-deleted assessments with pagination and optional filters.
      operationId: listAdminAssessments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: subject
          schema:
            $ref: '#/components/schemas/Subject'
        - in: query
          name: grade
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - in: query
          name: gradingType
          schema:
            $ref: '#/components/schemas/GradingType'
        - in: query
          name: scope
          description: Filter assessment ownership surface. Teachers only receive assessments they created.
          schema:
            type: string
            enum:
              - all
              - public
              - course
              - unplaced
        - in: query
          name: courseId
          description: Required when scope is course. Limits results to assessments placed on the selected course or one of its lessons.
          schema:
            type: string
            format: uuid
        - in: query
          name: visibility
          schema:
            $ref: '#/components/schemas/AssessmentVisibility'
      responses:
        '200':
          description: Assessment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAdminAssessmentsSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Assessments
      summary: Create assessment draft
      description: Creates a draft assessment. Exam assessments require a ready PDF source media before publish.
      operationId: createAssessment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
            example:
              title: De luyen tap toan 12
              subject: math
              grade: 12
              type: quiz
              gradingType: auto
              timeLimitMinutes: 90
              sourceMediaId: null
      responses:
        '201':
          description: Assessment draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAssessmentSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/assessments/{assessmentId}:
    get:
      tags:
        - Assessments
      summary: Get admin assessment detail
      operationId: getAdminAssessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      responses:
        '200':
          description: Assessment detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAssessmentDetailSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Assessments
      summary: Update assessment draft
      description: Only draft assessments can be updated.
      operationId: updateAssessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssessmentRequest'
      responses:
        '200':
          description: Assessment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAssessmentSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /admin/assessments/{assessmentId}/publish:
    post:
      tags:
        - Assessments
      summary: Publish assessment
      operationId: publishAssessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      responses:
        '200':
          description: Assessment published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAssessmentSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/assessments/{assessmentId}/clone:
    post:
      tags:
        - Assessments
      summary: Clone assessment
      description: Creates a new draft owned by the current actor. Placements and submissions are not cloned.
      operationId: cloneAssessment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloneAssessmentRequest'
      responses:
        '201':
          description: Assessment cloned as draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAssessmentSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/assessment-submissions/grading:
    get:
      tags:
        - Assessments
      summary: List submissions waiting for manual grading
      operationId: listGradingSubmissions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: assessmentId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Grading queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListGradingSubmissionsSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/assessment-submissions/{submissionId}:
    get:
      tags:
        - Assessments
      summary: Get grading submission detail
      operationId: getGradingSubmission
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubmissionIdPath'
      responses:
        '200':
          description: Grading submission detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGradingSubmissionDetailSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/assessment-submissions/{submissionId}/finalize:
    post:
      tags:
        - Assessments
      summary: Finalize manual or mixed grading
      operationId: finalizeManualSubmission
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubmissionIdPath'
      responses:
        '200':
          description: Submission finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSubmissionSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  parameters:
...
    AssessmentIdPath:
      in: path
      name: assessmentId
      required: true
      schema:
        type: string
        format: uuid
...
    SubmissionIdPath:
      in: path
      name: submissionId
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: BAD_REQUEST
              message: Validation failed
              details:
                - location: body
                  field: email
                  message: Invalid email
    Unauthorized:
      description: Missing, expired, or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated user does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with the current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    OffsetPagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 1015
        totalPages:
          type: integer
          example: 51
...
    Subject:
      type: string
      enum:
        - math
        - physics
        - chemistry
        - literature
        - english
        - biology
        - history
        - geography
    AssessmentType:
      type: string
      enum:
        - quiz
        - exam
    GradingType:
      type: string
      enum:
        - auto
        - manual
        - mixed
...
    AssessmentVisibility:
      type: string
      enum:
        - draft
        - published
        - hidden
...
    CreateAssessmentRequest:
      type: object
      required:
        - title
        - subject
        - grade
        - type
        - gradingType
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 255
        subject:
          $ref: '#/components/schemas/Subject'
        grade:
          type: integer
          minimum: 1
          maximum: 12
        gradingType:
          $ref: '#/components/schemas/GradingType'
        timeLimitMinutes:
          type: integer
          nullable: true
          minimum: 1
        sourceMediaId:
          type: string
          format: uuid
          nullable: true
    UpdateAssessmentRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 255
        subject:
          $ref: '#/components/schemas/Subject'
        grade:
          type: integer
          minimum: 1
          maximum: 12
        type:
          $ref: '#/components/schemas/AssessmentType'
        gradingType:
          $ref: '#/components/schemas/GradingType'
        timeLimitMinutes:
          type: integer
          nullable: true
          minimum: 1
        sourceMediaId:
          type: string
          format: uuid
          nullable: true
...
    AssessmentRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subject:
          $ref: '#/components/schemas/Subject'
        grade:
          type: integer
        type:
          $ref: '#/components/schemas/AssessmentType'
        gradingType:
          $ref: '#/components/schemas/GradingType'
        visibility:
          $ref: '#/components/schemas/AssessmentVisibility'
        createdById:
          type: string
          format: uuid
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        hiddenAt:
          type: string
          format: date-time
          nullable: true
        timeLimitMinutes:
          type: integer
          nullable: true
        sourceMediaId:
          type: string
          format: uuid
          nullable: true
        sourceMetadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAssessmentSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AssessmentRecord'
    AssessmentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subject:
          $ref: '#/components/schemas/Subject'
        grade:
          type: integer
        type:
          $ref: '#/components/schemas/AssessmentType'
        gradingType:
          $ref: '#/components/schemas/GradingType'
        visibility:
          $ref: '#/components/schemas/AssessmentVisibility'
        createdById:
          type: string
          format: uuid
          nullable: true
    AdminAssessmentPlacement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AssessmentPlacementType'
        slug:
          type: string
          nullable: true
        courseId:
          type: string
          format: uuid
          nullable: true
        lessonId:
          type: string
          format: uuid
          nullable: true
        isFeatured:
          type: boolean
        openTime:
          type: string
          format: date-time
          nullable: true
        closeTime:
          type: string
          format: date-time
          nullable: true
        maxAttempts:
          type: integer
          nullable: true
    AdminAssessmentListItem:
      allOf:
        - $ref: '#/components/schemas/AssessmentSummary'
        - type: object
          properties:
            timeLimitMinutes:
              type: integer
              nullable: true
            itemCount:
              type: integer
            submissionCount:
              type: integer
            createdById:
              type: string
              format: uuid
              nullable: true
            placements:
              type: array
              items:
                $ref: '#/components/schemas/AdminAssessmentPlacement'
            createdAt:
              type: string
              format: date-time
    ListAdminAssessmentsResponseData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminAssessmentListItem'
        pagination:
          $ref: '#/components/schemas/OffsetPagination'
    ListAdminAssessmentsSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ListAdminAssessmentsResponseData'
...
    AdminAssessmentItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderIndex:
          type: integer
        questionNumber:
          type: integer
          description: Computed from the item position within its section; not stored in the database.
        itemType:
          $ref: '#/components/schemas/AssessmentItemType'
        topicId:
          type: string
          format: uuid
          nullable: true
        topicName:
          type: string
          nullable: true
        difficulty:
          allOf:
            - $ref: '#/components/schemas/QuestionDifficulty'
          nullable: true
        maxScore:
          type: string
          example: '1.00'
        scoringConfig:
          nullable: true
        correctAnswer:
          nullable: true
        explanation:
          nullable: true
        question:
          $ref: '#/components/schemas/AdminAssessmentQuestion'
...
    AdminAssessmentSection:
      allOf:
        - $ref: '#/components/schemas/AssessmentSection'
        - type: object
          properties:
            sections:
              type: array
              items:
                $ref: '#/components/schemas/AdminAssessmentSection'
    AdminAssessmentDetail:
      allOf:
        - $ref: '#/components/schemas/AssessmentSummary'
        - type: object
          properties:
            timeLimitMinutes:
              type: integer
              nullable: true
            sourceMediaId:
              type: string
              format: uuid
              nullable: true
            sourceMedia:
              type: object
              additionalProperties: true
              nullable: true
            createdById:
              type: string
              format: uuid
              nullable: true
            submissionCount:
              type: integer
            placements:
              type: array
              items:
                $ref: '#/components/schemas/AdminAssessmentPlacement'
            sections:
              type: array
              items:
                $ref: '#/components/schemas/AdminAssessmentSection'
    AdminAssessmentDetailSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminAssessmentDetail'
...
    CloneAssessmentRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
...
    AssessmentSubmission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assessmentId:
          type: string
          format: uuid
        placementId:
          type: string
          format: uuid
          nullable: true
        attemptNumber:
          type: integer
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        startTime:
          type: string
          format: date-time
        submitTime:
          type: string
          format: date-time
          nullable: true
        autoScore:
          type: string
          nullable: true
        finalScore:
          type: string
          nullable: true
        violationCount:
          type: integer
          minimum: 0
        answers:
          type: object
          properties:
            mcq:
              type: array
              items:
                type: object
                properties:
                  itemId:
                    type: string
                    format: uuid
                  selectedOptionIds:
                    type: array
                    items:
                      type: string
                      format: uuid
            trueFalse:
              type: array
              items:
                type: object
                properties:
                  itemId:
                    type: string
                    format: uuid
                  optionId:
                    type: string
                    format: uuid
                  selectedValue:
                    type: boolean
            numeric:
              type: array
              items:
                type: object
                properties:
                  itemId:
                    type: string
                    format: uuid
                  answerValue:
                    type: string
            essay:
              type: array
              items:
                type: object
                properties:
                  itemId:
                    type: string
                    format: uuid
                  answer:
                    type: string
...
    AssessmentSubmissionSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AssessmentSubmission'
...
    ListGradingSubmissionsResponseData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminGradingSubmissionSummary'
        pagination:
          $ref: '#/components/schemas/OffsetPagination'
    ListGradingSubmissionsSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ListGradingSubmissionsResponseData'
    AdminGradingSubmissionDetail:
      allOf:
        - $ref: '#/components/schemas/AdminGradingSubmissionSummary'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AdminAssessmentItem'
            essayAnswers:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  itemId:
                    type: string
                    format: uuid
                  answer:
                    type: string
                  teacherScore:
                    type: string
                    nullable: true
                  teacherNote:
                    type: string
                    nullable: true
                  gradedAt:
                    type: string
                    format: date-time
                    nullable: true
    AdminGradingSubmissionDetailSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminGradingSubmissionDetail'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true